# service/common-job/docs/cron/openapi.yaml
openapi: 3.1.0
info:
  title: VSPO Common Cron
  description: This is the Cron documentation for VSPO Common services.
  contact: {}
  version: "1.0"
servers:
- url: http://localhost:8080
paths:
  /api/cron/creators:
    get:
      tags:
        - Channels
      summary: Upsert Channel(Youtube/Twitch/Twitcasting)
      description: Creates creators by fetching from Youtube using provided Channel IDs.
      parameters:
        - name: platform_type
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              enum: [youtube, twitch, twitcasting, niconico]
          style: form
          explode: false
          description: Comma-separated list of video platform types.
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [day, week, month, year]
          description: Period for performing updates.
        - name: creator_type
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              enum: [vspo, general]
          style: form
          explode: false
          description: Comma-separated list of creator types.
      responses:
        "200":
          description: Channels created successfully
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
      # security:
      #   - bearerAuth: []
  /api/cron/search_videos:
    get:
      tags:
        - Videos
      summary: Upsert videos
      description: Update videos related to a specific creator based on provided
      parameters:
        - name: platform_type
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              enum: [youtube, twitch, twitcasting, niconico]
          style: form
          explode: false
          description: Comma-separated list of video platform types.
        - name: video_type
          in: query
          required: true
          schema:
            type: string
            enum: [vspo_stream, clip, freechat]
          description: Type of the cron
      responses:
        "200":
          description: Videos updated successfully
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
      # security:
      #   - bearerAuth: []
  /api/cron/exist_videos:
    get:
      tags:
        - Videos
      summary: update exist videos
      description: Update exist videos
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [day, month, week]
          description: Period for performing updates.
      responses:
        "200":
          description: Videos updated successfully
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
      # security:
      #   - bearerAuth: []       
  /api/cron/channels:
    get:
      tags:
        - Channels
      summary: Update Channels
      description: Update channels
      parameters:
        - name: platform_type
          in: query
          required: true
          schema:
            type: string
            enum: [youtube, twitch, twitcasting, niconico]
      responses:
        "200":
          description: Videos updated successfully
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
      # security:
      #   - bearerAuth: []
  /api/ping:
    get:
      summary: Ping endpoint
      description: Returns a 200 status code if successful, or an error.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Pong
      # security:
      #   - bearerAuth: []
  /ping:
    post:
      summary: Ping endpoint
      description: Returns a 200 status code if successful, or an error.
      operationId: ping
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Pong
      # security:
      #   - bearerAuth: []
  /search_videos:
    post:
      tags:
        - Videos
      summary: Upsert videos
      description: Update videos related to a specific creator based on provided
      operationId: search_videos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platform_type:
                  type: array
                  items:
                    type: string
                    enum: [youtube, twitch, twitcasting, niconico]
                  description: List of video platform types.
                video_type:
                  type: string
                  enum: [vspo_stream, clip, freechat]
                  description: Type of the cron
      responses:
        "200":
          description: Videos updated successfully
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
  /exist_videos:
    post:
      tags:
        - Videos
      summary: Update exist videos
      description: Update exist videos based on the provided period.
      operationId: exist_videos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                period:
                  type: string
                  enum: [day, month, week]
                  description: Period for performing updates.
      responses:
        "200":
          description: Videos updated successfully
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CronResponse'


components:
  # securitySchemes:
  #   bearerAuth:
  #     type: http
  #     scheme: bearer
  #     bearerFormat: Bearer
  schemas:
    CronResponse:
      type: object
      properties:
        message:
          type: string
          example: success
# security:
#   - bearerAuth: []

tags:
  - name: Channels
    description: Operations related to creators
  - name: Videos
    description: Operations related to videos